Revert move of wasm to shared (for now), update extension webpack config
