+++ apps/namada-interface/package.json
@@ -15,7 +15,7 @@
     "@anoma/rpc": "0.1.0",
     "@anoma/seed-management": "0.1.0",
     "@anoma/session": "0.1.0",
-    "@anoma/shared": "0.1.0",
+    "@anoma/wasm": "0.1.0",
     "@anoma/tx": "0.1.0",
     "@anoma/utils": "0.1.0",
     "@anoma/wallet": "0.1.0",
