+++ packages/tx/src/Account.ts
@@ -1,4 +1,4 @@
-import { AnomaClient } from "@anoma/shared";
+import { AnomaClient } from "@anoma/wasm";
 import { Tokens } from "@anoma/tx";
 
 class Account {
