+++ packages/wasm/lib/.appveyor.yml
