+++ packages/wasm/lib/src/types/encrypted_keypair.rs
