+++ packages/wasm/lib/src/types/mnemonic.rs
