+++ packages/wasm/lib/src/utils.rs
