+++ packages/wasm/lib/src/wallet.rs
